requirements.txt
fastapi 
aioredis